+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.3] - 2022-01-26
+### Deployed
+- Deploy a new `DelegatableVaultMaker` with `withdrawToken()` on optimism
+
+## [1.3.3-staging] - 2022-01-24
+### Deployed
+- Deploy a new `DelegatableVaultMaker` with `withdrawToken()` on optimismKovan
+
 ## [1.3.2-staging] - 2022-01-07
 - Deploy a new DelegatableVaultMaker on **Optimism Kovan** based on `@perp/curie-contract#1.0.7`